requirements.txt
fastapi
uvicorn
pydantic[email]
supabase
bcrypt
google-api-python-client
google-auth
